requirements.txt
fastapi
uvicorn
passlib[bcrypt]
python-multipart
psycopg2-binary
redis